sub_redis_init/requirements.txt
annotated-types==0.7.0
async-timeout==4.0.3
certifi==2024.6.2
charset-normalizer==3.3.2
idna==3.7
orjson==3.9.14
pydantic==2.6.1
pydantic-settings==2.2.1
pydantic_core==2.16.2
python-dotenv==1.0.0
redis==5.0.1
requests==2.32.3
typing_extensions==4.12.1
urllib3==2.2.1
